requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
msgpack
grpcio
protobuf
matplotlib
